% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.suquan.R
\name{predict.suquan}
\alias{predict.suquan}
\title{Predict with a supervised quantiled linear model.}
\usage{
\method{predict}{suquan}(m, x)
}
\arguments{
\item{m}{The model, a list that contains three fields: \code{f}, the quantile
function, \code{b}, the linear model, \code{b0}, the intercept}

\item{x}{Samples (one per row) for which the prediction is made.}
}
\value{
A vector of predictions, for all samples.
}
\description{
Make prediction on new samples with a quantiles linear model.
}
\examples{
n <- 100
p <- 40
x <- matrix(rnorm(n*p),n,p)
y <- rbinom(n,1,0.5)*2-1
lambda <- 0.2*max(abs(crossprod(y,x)))/n
m <- suquan(x, y, family="gaussian", penalty="elasticnet", opts=list(alpha=0))
predict.suquan(m,x)
}

