% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/suquan.R
\name{suquan}
\alias{suquan}
\title{Fit a supervised quantiled linear model.}
\usage{
suquan(x, y, family = c("gaussian", "binomial"), penalty = "elasticnet",
  lambda = 1, intercept = TRUE, f_init = NULL, maxiter = 10,
  eps = 1e-06, use.glmnet = TRUE, opts = list())
}
\arguments{
\item{x}{The input matrix, each row is a sample, each column a feature.}

\item{y}{The response variable. Quantitative for \code{family="gaussian"},
binary with values \code{+1} and \code{-1} for \code{family="binomial"}}

\item{family}{The response type. For \code{family="gaussian"}, ...}

\item{penalty}{The penalty type.}

\item{lambda}{The scaling of the penalty (default \code{1})}

\item{intercept}{Should intercept(s) be fitted (default=\code{TRUE}) or set
to zero (\code{FALSE})}

\item{f_init}{The initial distribution for the quantile transformation
(default is a Gaussian cdf)}

\item{maxiter}{Maximum number of times the optimization loop (once in the
quantile distribution and once in the model) is performed (default
\code{10})}

\item{eps}{Stopping criterion: the loop will stop when the norm of the
difference between the quantile distribution after two successive
optimization is smaller than \code{eps} (default \code{1e-6})}

\item{use.glmnet}{Whether the \code{glmnet} package should be used to fit the
  model for a given quantile distribution. If \code{FALSE}, then the less
  optimized \code{apg} optimizer is used (default \code{TRUE}).}

\item{opts}{List of parameters, which must include: \itemize{ \item }}
}
\value{
toto
}
\description{
Fit a generalized linear model via penalized maximum likelihood, with joint
optimization of full quantile normalization.
}
\examples{
n <- 100
p <- 40
x <- matrix(rnorm(n*p),n,p)
y <- rbinom(n,1,0.5)*2-1
lambda <- 0.2*max(abs(crossprod(y,x)))/n
m <- suquan(x, y, family="gaussian", penalty="elasticnet", opts=list(alpha=0))
}

